dearseq: Differential Expression Analysis for RNA-seq data through a robust variance component test

Differential Expression Analysis RNA-seq data with variance component score test accounting for data heteroscedasticity through precision weights. Perform both gene-wise and gene set analyses, and can deal with repeated or longitudinal data. Methods are detailed in: i) Agniel D & Hejblum BP (2017) Variance component score test for time-course gene set analysis of longitudinal RNA-seq data, Biostatistics, 18(4):589-604 ; and ii) Gauthier M, Agniel D, Thiébaut R & Hejblum BP (2020) dearseq: a variance component score test for RNA-Seq differential analysis that effectively controls the false discovery rate, NAR Genomics and Bioinformatics, 2(4):lqaa093.

Details

Analysis of RNA-seq data with variance component score test accounting for data heteroscedasticity through precision weights. Performs gene-wise analysis as well as gene set analysis, including for complex experimental designs such as longitudinal data.

Package:	dearseq

Type:	Package

Version:	1.13.4

Date:	2023-09-06

License:	GPL-2

The two main functions of the dearseq package are dear_seq and dgsa_seq.

References

Agniel D & Hejblum BP (2017). Variance component score test for time-course gene set analysis of longitudinal RNA-seq data, Biostatistics, 18(4):589-604. DOI: 10.1093/biostatistics/kxx005. arXiv:1605.02351.

Gauthier M, Agniel D, Thiébaut R & Hejblum BP (2020). dearseq: a variance component score test for RNA-Seq differential analysis that effectively controls the false discovery rate, NAR Genomics and Bioinformatics, 2(4):lqaa093. DOI: 10.1093/nargab/lqaa093. DOI: 10.1101/635714

See also

Author

Maintainer: Boris P. Hejblum boris.hejblum@u-bordeaux.fr (ORCID)

Authors:

Denis Agniel denis.agniel@gmail.com

Marine Gauthier marine.gauthier@u-bordeaux.fr

Other contributors:

Mélanie Huchon melanie.huchon@u-bordeaux.fr [contributor]

Iban Larrouture iban.larrouture@u-bordeaux.fr [contributor]